+++ datanode/gateway/graphql/resolvers.go
@@ -122,13 +122,27 @@ func (r *VegaResolverRoot) MarginLevels() MarginLevelsResolver {
 }
 
 // MarginLevels returns the market levels resolver.
-func (r *VegaResolverRoot) AbstractMarginLevels() AbstractMarginLevelsResolver {
-	return (*myAbstractMarginLevelsResolver)(r)
+func (r *VegaResolverRoot) MarginEstimate() MarginEstimateResolver {
+	return (*myMarginEstimateResolver)(r)
 }
 
 // MarginLevelsUpdate returns the market levels resolver.
 func (r *VegaResolverRoot) MarginLevelsUpdate() MarginLevelsUpdateResolver {
 	return (*myMarginLevelsUpdateResolver)(r)
+}
+
+func (r *VegaResolverRoot) OrderInfo() OrderInfoResolver {
+	return (*myOrderInfoResolver)(r)
+}
+
+type myOrderInfoResolver VegaResolverRoot
+
+func (r *myOrderInfoResolver) Remaining(ctx context.Context, obj *v2.OrderInfo, data string) error {
+	remaining, err := strconv.ParseUint(data, 10, 64)
+	if err == nil {
+		obj.Remaining = remaining
+	}
+	return err
 }
 
 // PriceLevel returns the price levels resolver.
@@ -1320,7 +1334,7 @@ func (r *myQueryResolver) EstimatePosition(
 	marketId string,
 	openVolume string,
 	averageEntryPrice string,
-	orders []*OrderInfo,
+	orders []*v2.OrderInfo,
 	marginAccountBalance string,
 	generalAccountBalance string,
 	orderMarginAccountBalance string,
@@ -1334,26 +1348,11 @@ func (r *myQueryResolver) EstimatePosition(
 		return nil, err
 	}
 
-	ord := make([]*v2.OrderInfo, 0, len(orders))
-	for _, o := range orders {
-		r, err := safeStringUint64(o.Remaining)
-		if err != nil {
-			return nil, err
-		}
-
-		ord = append(ord, &v2.OrderInfo{
-			Side:          o.Side,
-			Price:         o.Price,
-			Remaining:     r,
-			IsMarketOrder: o.IsMarketOrder,
-		})
-	}
-
 	req := &v2.EstimatePositionRequest{
 		MarketId:                  marketId,
 		OpenVolume:                ov,
 		AverageEntryPrice:         averageEntryPrice,
-		Orders:                    ord,
+		Orders:                    orders,
 		MarginAccountBalance:      marginAccountBalance,
 		GeneralAccountBalance:     generalAccountBalance,
 		OrderMarginAccountBalance: orderMarginAccountBalance,
@@ -2519,44 +2518,6 @@ func (r *myMarginLevelsUpdateResolver) OrderMarginLevel(_ context.Context, m *ve
 }
 
 func (r *myMarginLevelsUpdateResolver) MarginFactor(_ context.Context, m *vegapb.MarginLevels) (string, error) {
-	return m.MarginFactor, nil
-}
-
-// BEGIN: AbstractMarginLevels Resolver
-
-type myAbstractMarginLevelsResolver VegaResolverRoot
-
-// END: AbstractMarginLevels Resolver
-
-func (r *myAbstractMarginLevelsResolver) Market(ctx context.Context, m *vegapb.MarginLevels) (*vegapb.Market, error) {
-	return r.r.getMarketByID(ctx, m.MarketId)
-}
-
-func (r *myAbstractMarginLevelsResolver) Asset(ctx context.Context, m *vegapb.MarginLevels) (*vegapb.Asset, error) {
-	return r.r.getAssetByID(ctx, m.Asset)
-}
-
-func (r *myAbstractMarginLevelsResolver) CollateralReleaseLevel(_ context.Context, m *vegapb.MarginLevels) (string, error) {
-	return m.CollateralReleaseLevel, nil
-}
-
-func (r *myAbstractMarginLevelsResolver) InitialLevel(_ context.Context, m *vegapb.MarginLevels) (string, error) {
-	return m.InitialMargin, nil
-}
-
-func (r *myAbstractMarginLevelsResolver) SearchLevel(_ context.Context, m *vegapb.MarginLevels) (string, error) {
-	return m.SearchLevel, nil
-}
-
-func (r *myAbstractMarginLevelsResolver) MaintenanceLevel(_ context.Context, m *vegapb.MarginLevels) (string, error) {
-	return m.MaintenanceMargin, nil
-}
-
-func (r *myAbstractMarginLevelsResolver) OrderMarginLevel(_ context.Context, m *vegapb.MarginLevels) (string, error) {
-	return m.OrderMargin, nil
-}
-
-func (r *myAbstractMarginLevelsResolver) MarginFactor(_ context.Context, m *vegapb.MarginLevels) (string, error) {
 	return m.MarginFactor, nil
 }